Wed, Jan 5, 2022, 8:21 AM +00:00

+++ volume/volume.c
@@ -4,7 +4,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-typedef int16_t BYTE;
+typedef uint8_t BYTE;
+//typedef int16_t BYTE;
 
 // Number of bytes in .wav header
 const int HEADER_SIZE = 44;
